# Conflicts:
#	compiler/shared/main/scala/mlscript/compiler/codegen/CppAst.scala
#	compiler/shared/main/scala/mlscript/compiler/codegen/CppCodeGen.scala
#	compiler/shared/main/scala/mlscript/compiler/ir/IR.scala
#	compiler/shared/test/scala/mlscript/compiler/TestIR.scala

+++ compiler/shared/main/scala/mlscript/compiler/codegen/CppAst.scala
@@ -122,7 +122,7 @@ object Expr:
 enum Expr:
   case Var(name: Str)
   case IntLit(value: BigInt)
-  case DoubleLit(value: Float)
+  case DoubleLit(value: Double)
   case StrLit(value: Str)
   case CharLit(value: Char)
   case Call(func: Expr, args: Ls[Expr])
